week_05/more_APIs/requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
oauthlib==3.0.1
PySocks==1.6.8
python-crontab==2.3.6
python-dateutil==2.8.0
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
slackclient==1.3.1
tweepy==3.7.0
urllib3==1.24.1
websocket-client==0.54.0
